fix(root): fix btn on keyboard open, fix locales, fix ttn and amount on backclick, fix scroll in list
