+++ src/i18n/locales/en.json
@@ -146,7 +146,7 @@
     "detailsDimensions": "Dimensions",
     "detailsButton": "Continue",
     "createEmptyDescription": "Specify the amount the recipient must pay. The fee will be deducted upon receipt — you will receive the amount minus the fee after the delivery is confirmed.",
-    "createFullfiledDescription": "The recipient will pay this amount. After deducting the fee, you will receive approximately {amount− ${{fee}}}.`}",
+    "createFullfiledDescription": "The recipient will pay this amount. After deducting the fee, you will receive approximately {amount− ${{ fee }}}.",
     "createButton": "Create Deal",
     "sendDescription": "Share this link with the byer",
     "sendSubtitle": "Package Payment",
