+++ src/i18n/locales/uk.json
@@ -145,7 +145,7 @@
     "detailsDimensions": "Габарити",
     "detailsButton": "Продовжити",
     "createEmptyDescription": "Вкажіть суму, яку має оплатити одержувач. Комісію буде утримано при отриманні — ви отримаєте суму за вирахуванням комісії після підтвердження доставки.",
-    "createFullfiledDescription": "Одержувач сплатить цю суму. Після вирахування комісії ви отримаєте приблизно {amount− ${{fee}}}.",
+    "createFullfiledDescription": "Одержувач сплатить цю суму. Після вирахування комісії ви отримаєте приблизно {сумма− ${{fee}}}.",
     "createButton": "Створити угоду",
     "sendDescription": "Поділіться цим посиланням з покупцем",
     "sendSubtitle": "Оплата посилки",
