+++ src/pages/SendPackage/SendPackage.tsx
@@ -1,63 +1,15 @@
+import { Button } from "../../components/Button";
 import { ESendPakageTab } from "./SendPackage.type";
 import { NavHeader } from "../../components/NavHeader";
-import { SendPackageDetails } from "../../components/SendPackageDetails";
-import { SendPackagePaymentCreate } from "../../components/SendPackagePaymentCreate";
-import { SendPackagePaymentSend } from "../../components/SendPackagePaymentSend";
-import { SendPackageValidateTTN } from "../../components/SendPackageValidateTTN";
 import classNames from "classnames";
+import loader from "../../assets/images/loader.webp";
 import styles from "./SendPackage.module.css";
-import { t } from "i18next";
 import { useKeyboardStatus } from "../../hooks/useKeyboardStatus";
-import { usePackageStore } from "../../store/packageStore";
+import { useSendPackage } from "../../hooks/useSendPackage";
 
 const SendPackage = () => {
-  const { data, setActiveTab } = usePackageStore();
+  const { btnData, backData, titleTab, data, switchTab } = useSendPackage();
   const { isKeyboardOpen } = useKeyboardStatus();
-
-  const tabTitle = {
-    [ESendPakageTab.ValidateTTN]: t("sendPackage.validateTitle"),
-    [ESendPakageTab.PackageDetails]: t("sendPackage.detailsTitle"),
-    [ESendPakageTab.PaymentCreate]: t("sendPackage.createTitle"),
-    [ESendPakageTab.PaymentSend]: ""
-  };
-
-  const switchTab = () => {
-    switch (data.activeTab) {
-      case ESendPakageTab.ValidateTTN:
-        return <SendPackageValidateTTN />;
-      case ESendPakageTab.PackageDetails:
-        return <SendPackageDetails />;
-      case ESendPakageTab.PaymentCreate:
-        return <SendPackagePaymentCreate />;
-      case ESendPakageTab.PaymentSend:
-        return <SendPackagePaymentSend />;
-      default:
-        return <SendPackageValidateTTN />;
-    }
-  };
-
-  const handleBack = {
-    [ESendPakageTab.ValidateTTN]: {
-      isLink: true,
-      link: "/",
-      action: undefined
-    },
-    [ESendPakageTab.PackageDetails]: {
-      isLink: false,
-      link: undefined,
-      action: () => setActiveTab(ESendPakageTab.ValidateTTN)
-    },
-    [ESendPakageTab.PaymentCreate]: {
-      isLink: false,
-      link: undefined,
-      action: () => setActiveTab(ESendPakageTab.PackageDetails)
-    },
-    [ESendPakageTab.PaymentSend]: {
-      isLink: false,
-      link: undefined,
-      action: () => setActiveTab(ESendPakageTab.PaymentCreate)
-    }
-  };
 
   return (
     <main
@@ -69,16 +21,36 @@ const SendPackage = () => {
         <div className="flex h-full flex-col">
           <div className="mb-[30px] mt-5 px-[1rem]">
             <NavHeader
-              isLink={handleBack[data.activeTab].isLink}
-              link={handleBack[data.activeTab].link}
-              action={handleBack[data.activeTab].action}
+              isLink={backData.isLink}
+              link={backData.link}
+              action={backData.action}
             />
           </div>
           <div className={styles.tab__box}>
             {data.activeTab !== ESendPakageTab.PaymentSend && (
-              <p className={styles.tab__title}>{tabTitle[data.activeTab]}</p>
+              <p className={styles.tab__title}>{titleTab}</p>
             )}
             {switchTab()}
+          </div>
+          <div className="custom-container primary-button-container">
+            <Button
+              actionHandler={btnData.action}
+              disabled={btnData.disabled}
+              className={classNames({
+                "!translate-y-[80px]": isKeyboardOpen
+              })}
+            >
+              {btnData.loading ? (
+                <img
+                  className="spinner"
+                  src={loader}
+                  width={"24px"}
+                  height={"24px"}
+                />
+              ) : (
+                <p>{btnData.title}</p>
+              )}
+            </Button>
           </div>
         </div>
       </div>
